feat(git): improve the `Git` interface
### General ideas

Push and pull operations are remote operations that should live under remotes. Tag pushing is the same but it is tightly tied to tags, so leaving it there.

I want to balance the API depth by cleaning up ungrouped operations, and bringing down operations that are three levels deep. Still, `git init` needs to be at the top-level, to avoid `repo.repo.init()`.

### Detailed changes

- move `repo.commits.push()` to `repo.remotes.push()`
- move `repo.commits.pull()` to `repo.remotes.pull()`
- move `repo.clone()` to `repo.remotes.clone()`
- move `repo.branches.upstream.set()` to `repo.branches.setUpstream()`
- move `repo.branches.upstream.unset()` to `repo.branches.unsetUpstream()`
- split `repo.ignore.check()` to `repo.ignore.ignored()` and `repo.ignore.unignored()`
- rename `repo.commits.head()` to `commits.current()`: head is a branch
- rename `checkout({ new })` to `checkout({ create })`

## Follow up

This is a good state to be, but the following changes will be made before the next release:

- `repo.commits.get()`
- `repo.init({ directory })`
- `repo.branches.switch()`
- `repo.commits.revert()`
- `repo.remotes.current()` or `git().remotes.default()`, instead of assuming "origin"
- `repo.branches.merge()`
- `repo.branches.rebase()`
- `repo.branches.squash()`