refactor(git)!: updated `Git` interface (#388)
### General ideas

Singular names for grouping, because this matches the git nomenclature
better. I was doing a lot of mental hopping with `commits`, `branches`
etc. The plural "collections" interface approach is a good practice, but
it's getting in the way.

Push and pull operations are remote operations that should live under
remotes. Tag pushing is the same but it is tightly tied to tags, so
leaving it there.

I want to balance the API depth by cleaning up ungrouped operations, and
bringing down operations that are three levels deep. Still, `git init`
needs to be at the top-level, to avoid `repo.repo.init()`.

Making getters return non-nullable, and throw error on exceptional
states. For example, there is no commit in empty repos, or there is no
current branch in detached head state. These however are rare cases and
forcing the user to check validity after common calls is a burden.

### Detailed changes

- move `git().commits.pull()` to `git().remote.pull()`
- move `git().commits.push()` to `git().remote.push()`
- move `git().clone()` to `git().remote.clone()`
- move `git().branches.upstream.set()` to `git().branch.track()`
- move `git().branches.upstream.unset()` to `git().branches.untrack()`
- split `git().ignore.check()` to `repo.ignore.filter()` and
`git().ignore.omit()`
- rename `git().branch.checkout({ new })` to `git().branch.checkout({
create })`
- rename `git().remotes.defaultBranch()` to `git().remote.head()`
- move `.remotes.*` to `git().remote.*`
- move `.branches.*` to `git().branch.*`
- move `git().commits.*` to `git().commit*`
- move `git().tags.*` to `git().tag.*`
- `git().remote.head()` returns non-nullable and throws on detached head
- `git().branch.current()` returns non-nullable and throws on detached
head
- `git().commit.log()` returns empty array on empty repo

### Follow up

This is a good state to be, but the following additions will be made
before the next release:

- `git().init({ directory })`
- `git().config.get()`
- `git().remote.remove()`
- `git().remote.fetch()`
- `git().branch.switch()`
- `git().index.reset()`
- `git().commit.get()`
- `git().commit.revert()`
- `git().tag.delete()`
- `git().worktree.*`
- `git().stash.*`
- `git().merge.*`
- `git().rebase.*`